refactor: service interface로 변경
