+++ src/main/java/com/example/projectlxp/category/service/CategoryService.java
@@ -5,20 +5,9 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 
 import com.example.projectlxp.category.dto.CategoryDTO;
-import com.example.projectlxp.category.repository.CategoryRepository;
 
 @Service
-public class CategoryService {
+public interface CategoryService {
 
-    private final CategoryRepository categoryRepository;
-
-    public CategoryService(CategoryRepository categoryRepository) {
-        this.categoryRepository = categoryRepository;
-    }
-
-    public List<CategoryDTO> getCategories() {
-        return categoryRepository.findAllCategoryOptimize().stream()
-                .map(CategoryDTO::from)
-                .toList();
-    }
+    List<CategoryDTO> getCategories();
 }
